Move Game out of containers into a component
